Move Applications to own folder and rename to TrainingApplications
First, we'll be adding more related classes, second it's used as $trainingApplications anyway.
