+++ site/tests/Training/Applications/TrainingApplicationsTest.phpt
@@ -5,17 +5,18 @@ namespace MichalSpacekCz\Training;
 
 use MichalSpacekCz\ShouldNotHappenException;
 use MichalSpacekCz\Test\Database\Database;
+use MichalSpacekCz\Training\Applications\TrainingApplications;
 use Tester\Assert;
 use Tester\TestCase;
 
-$runner = require __DIR__ . '/../bootstrap.php';
+$runner = require __DIR__ . '/../../bootstrap.php';
 
 /** @testCase */
-class ApplicationsTest extends TestCase
+class TrainingApplicationsTest extends TestCase
 {
 
 	public function __construct(
-		private readonly Applications $applications,
+		private readonly TrainingApplications $trainingApplications,
 		private readonly Database $database,
 	) {
 	}
@@ -24,14 +25,14 @@ class ApplicationsTest extends TestCase
 	public function testGetValidUnpaidCount(): void
 	{
 		$this->database->setFetchFieldDefaultResult(909);
-		Assert::same(909, $this->applications->getValidUnpaidCount());
+		Assert::same(909, $this->trainingApplications->getValidUnpaidCount());
 
 		$this->database->setFetchFieldDefaultResult('\o/');
 		Assert::exception(function (): void {
-			$this->applications->getValidUnpaidCount();
+			$this->trainingApplications->getValidUnpaidCount();
 		}, ShouldNotHappenException::class, 'Count is a string not an integer');
 	}
 
 }
 
-$runner->run(ApplicationsTest::class);
+$runner->run(TrainingApplicationsTest::class);
